+++ examples/hacker-news/config.dev.replica.js
@@ -1,0 +1,42 @@
+import devConfig from './config.dev'
+
+const devReplicaConfig = {
+  ...devConfig,
+  name: 'hacker-news-replica',
+  port: '3001',
+  distDir: 'dist-replica',
+  readModelConnectors: {
+    default: {
+      module: '@resolve-js/readmodel-lite',
+      options: {
+        databaseFile: 'data-replica/read-model-default.db',
+      },
+    },
+    hackerNews: {
+      module: '@resolve-js/readmodel-lite',
+      options: {
+        databaseFile: 'data-replica/read-model-hackerNews.db',
+      },
+    },
+    comments: {
+      module: '@resolve-js/readmodel-lite',
+      options: {
+        databaseFile: 'data-replica/read-model-comments.db',
+      },
+    },
+    elasticSearch: {
+      module: 'common/read-models/elastic-search-connector.js',
+      options: {
+        /*
+        node: "<your-cloud-elastic-search-host>:port",
+        auth: {
+          username: 'name',
+          password: 'pass'
+        }
+        */
+      },
+    },
+  },
+}
+
+export default devReplicaConfig