+++ examples/shopping-list-tutorial/lesson-1/config.app.js
@@ -1,0 +1,14 @@
+const appConfig = {
+  aggregates: [
+    {
+      name: 'ShoppingList',
+      commands: 'common/aggregates/shopping_list.commands.js',
+      projection: 'common/aggregates/shopping_list.projection.js',
+    },
+  ],
+  readModels: [],
+  viewModels: [],
+  clientEntries: ['client/index.js'],
+}
+
+export default appConfig