+++ examples/shopping-list-tutorial/lesson-2/config.app.js
@@ -6,7 +6,14 @@ const appConfig = {
       projection: 'common/aggregates/shopping_list.projection.js',
     },
   ],
-  readModels: [],
+  readModels: [
+    {
+      name: 'ShoppingLists',
+      projection: 'common/read-models/shopping_lists.projection.js',
+      resolvers: 'common/read-models/shopping_lists.resolvers.js',
+      connectorName: 'default',
+    },
+  ],
   viewModels: [],
   clientEntries: ['client/index.js'],
 }