+++ examples/shopping-list-tutorial/lesson-5/common/view-models/shopping_list.projection.js
@@ -1,15 +1,24 @@
 import {
   SHOPPING_LIST_CREATED,
+  SHOPPING_LIST_REMOVED,
   SHOPPING_ITEM_CREATED,
   SHOPPING_ITEM_TOGGLED,
+  SHOPPING_ITEM_REMOVED,
 } from '../eventTypes'
 
 export default {
-  Init: () => null,
+  Init: () => ({
+    id: 'id',
+    name: 'unnamed',
+    list: [],
+  }),
   [SHOPPING_LIST_CREATED]: (state, { aggregateId, payload: { name } }) => ({
     id: aggregateId,
     name,
     list: [],
+  }),
+  [SHOPPING_LIST_REMOVED]: () => ({
+    removed: true,
   }),
   [SHOPPING_ITEM_CREATED]: (state, { payload: { id, text } }) => ({
     ...state,
@@ -32,5 +41,9 @@ export default {
           }
         : item
     ),
+  }),
+  [SHOPPING_ITEM_REMOVED]: (state, { payload: { id } }) => ({
+    ...state,
+    list: state.list.filter((item) => item.id !== id),
   }),
 }