+++ packages/modules/module-admin/src/commands/system.js
@@ -1,0 +1,3 @@
+export const command = 'system'
+export const describe = "manage an application's system"
+export const builder = (yargs) => yargs.commandDir('system')